requirements_extended.txt
# Dependências originais
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
opencv-python>=4.8.0
librosa>=0.10.0
sqlalchemy>=2.0.0
openai-whisper>=20230314
moviepy>=1.0.3
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Dependências para extensão com imagens e MongoDB
pymongo>=4.5.0  # MongoDB driver
pillow>=10.0.0  # Processamento de imagens adicional
requests>=2.31.0  # Para downloads de modelos se necessário

# Modelos específicos para análise de imagens
# Nota: Os modelos do HuggingFace serão baixados automaticamente:
# - nlpconnect/vit-gpt2-image-captioning (para legendas)
# - google/vit-base-patch16-224 (para classificação)
# - Salesforce/blip-image-captioning-base (alternativa)
